add f support

+++ src/ft_vsprintf.c
@@ -6,7 +6,7 @@
 /*   By: qle-guen <qle-guen@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/06/19 15:25:52 by qle-guen          #+#    #+#             */
-/*   Updated: 2016/06/19 15:26:54 by qle-guen         ###   ########.fr       */
+/*   Updated: 2016/06/19 17:14:33 by qle-guen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,5 +16,5 @@
 int				ft_vsprintf
 	(char *str, const char *format, va_list ap)
 {
-	return (ft_vsnprintf(str, INT_MAX + 1, format, ap));
+	return (ft_vsnprintf(str, (size_t)INT_MAX + 1, format, ap));
 }
